Much improved the Windows Phone version
Every TrackPageView call generated a fake new session and a fake new
visitor in Google Analytics, furthermore no screen resolution, OS
version, device manufacturer and model was sent.
I've written the class WindowsPhoneAnalyticsSession that take care to
automatically create the cookie (with the right informtion in it, now)
and session id, saving permanent information in the IsolatedStorage,
added the needed parameters for screen resolution and refactored
useragent to be the same that IE9 use.
Now every problem listed at the begin is correctly working.
Furthermore now the tracker automatically initialize itself UserAgent,
AnalyticsSession and Language, so you only need to create the
GoogleAnalyticsTracker object now and it's ready to go out of the box.