Use Headers.TryAddWithoutValidation for User-Agent header
.NET validates headers when you add them. See: https://stackoverflow.com/questions/13198090/adding-httpclient-headers-generates-a-formatexception-with-some-values
Normally this is nice, but in this library we are usually passing on the User-Agent header from the original request.
The problem with this is that .NET's validation may fail in cases that GoogleAnalytics would handle correctly.

When I enabled error logging of the Tracker, I got quite a number of these `System.FormatException`s in my log.

By using TryAddWithoutValidation we bypass this problematic validation.